Functions.php added, File Structure completed
Image Upload offline

+++ intern/sql_ini.php
@@ -1,7 +1,11 @@
 <?php session_start() ?>
-<?php include_once '../config.php'; ?>
-<?php include_once '../pages/html_prepare.php'; ?>
-<?php include_once '../pages/session_check.php'; ?>
+<?php include_once '../backend/config.php'; ?>
+<?php include_once '../backend/html_prepare.php'; ?>
+<?php include_once '../backend/session_check.php'; ?>
+<?php if ($_SESSION['user_role'] != '3') {
+  echo "Sie besitzen nicht die benötigten Rechte um auf diese Seite zuzugreifen";
+  die();
+} ?>
 <?php
 $statement = 'DROP DATABASE blurry';
 $pdo->exec($statement);
